backend/api/requirements.txt
fastapi==0.124.0
httpx==0.28.1
pyjwt==2.10.1
cryptography==44.0.0
msal